refactoring metrics visualization
