+++ tests/test_metrics_constraintplot.py
@@ -1,24 +1,14 @@
-import os, shutil
 import jijmodeling as jm
-from matplotlib import axes, figure
-
 import numpy as np
 import pandas as pd
 import pytest
+from matplotlib import axes, figure
 
 import jijbench as jb
-
-from jijbench.visualization.metrics.plot import _get_violations_dict, MetricsPlot
-
-
-@pytest.fixture(scope="function", autouse=True)
-def pre_post_process():
-    # preprocess
-    yield
-    # postprocess
-    norm_path = os.path.normcase("./.jb_results")
-    if os.path.exists(norm_path):
-        shutil.rmtree(norm_path)
+from jijbench.visualization.metrics.constraintplot.constraintplot import (
+    ConstraintPlot,
+    _get_violations_dict,
+)
 
 
 def solve():
@@ -76,261 +66,6 @@ def test_metrics_plot_get_violations_dict():
     assert (violations_dict["onehot_violations"] == np.array([0, 2])).all()
 
 
-def test_metrics_plot_boxplot_return_value():
-    num_multipliers = 2
-    params = {"multipliers": [{} for _ in range(num_multipliers)]}
-    solver_list = [solve]
-
-    expect_num_fig = len(params["multipliers"]) * len(solver_list)
-
-    bench = jb.Benchmark(
-        params=params,
-        solver=solver_list,
-    )
-    result = bench()
-    mplot = MetricsPlot(result)
-    fig_ax_tuple = mplot.boxplot(f=simple_func_for_boxplot)
-    assert len(fig_ax_tuple) == expect_num_fig
-    assert type(fig_ax_tuple[0][0]) == figure.Figure
-    assert type(fig_ax_tuple[0][1]) == axes.Subplot
-
-
-def test_metrics_plot_boxplot_call_maplotlib_boxplot(mocker):
-    m = mocker.patch("matplotlib.axes.Subplot.boxplot")
-
-    num_multipliers = 2
-    params = {"multipliers": [{} for _ in range(num_multipliers)]}
-    solver_list = [solve]
-
-    expect_num_call = len(params["multipliers"]) * len(solver_list)
-
-    bench = jb.Benchmark(
-        params=params,
-        solver=solver_list,
-    )
-    result = bench()
-    mplot = MetricsPlot(result)
-    mplot.boxplot(f=simple_func_for_boxplot)
-
-    assert m.call_count == expect_num_call
-
-
-def test_metrics_plot_boxplot_arg_figsize():
-    figwidth, figheight = 8, 4
-
-    bench = jb.Benchmark(
-        params={},
-        solver=[solve],
-    )
-    result = bench()
-    mplot = MetricsPlot(result)
-    fig_ax_tuple = mplot.boxplot(
-        f=simple_func_for_boxplot,
-        figsize=(figwidth, figheight),
-    )
-    fig, ax = fig_ax_tuple[0]
-
-    assert fig.get_figwidth() == 8
-    assert fig.get_figheight() == 4
-
-
-def test_metrics_plot_boxplot_arg_title(mocker):
-    m = mocker.patch("matplotlib.figure.Figure.suptitle")
-
-    title = ["title1", "title2"]
-
-    num_multipliers = 2
-    params = {"multipliers": [{} for _ in range(num_multipliers)]}
-    solver_list = [solve]
-
-    bench = jb.Benchmark(
-        params=params,
-        solver=solver_list,
-    )
-    result = bench()
-    mplot = MetricsPlot(result)
-    mplot.boxplot(
-        f=simple_func_for_boxplot,
-        title=title,
-    )
-    m.assert_has_calls(
-        [
-            mocker.call("title1", fontsize=None),
-            mocker.call("title2", fontsize=None),
-        ]
-    )
-
-
-def test_metrics_plot_boxplot_arg_title_fontsize(mocker):
-    m = mocker.patch("matplotlib.figure.Figure.suptitle")
-
-    title = ["title1", "title2"]
-    title_fontsize = 12
-
-    num_multipliers = 2
-    params = {"multipliers": [{} for _ in range(num_multipliers)]}
-    solver_list = [solve]
-
-    bench = jb.Benchmark(
-        params=params,
-        solver=solver_list,
-    )
-    result = bench()
-    mplot = MetricsPlot(result)
-    mplot.boxplot(
-        f=simple_func_for_boxplot,
-        title=title,
-        title_fontsize=title_fontsize,
-    )
-    m.assert_has_calls(
-        [
-            mocker.call("title1", fontsize=title_fontsize),
-            mocker.call("title2", fontsize=title_fontsize),
-        ]
-    )
-
-
-def test_metrics_plot_boxplot_arg_xticklabels():
-    bench = jb.Benchmark(
-        params={},
-        solver=[solve],
-    )
-    result = bench()
-    mplot = MetricsPlot(result)
-    fig_ax_tuple = mplot.boxplot(
-        f=simple_func_for_boxplot,
-    )
-    ax = fig_ax_tuple[0][1]
-    xticklabels = ax.get_xticklabels()
-
-    assert xticklabels[0].get_text() == "data1"
-    assert xticklabels[1].get_text() == "data2"
-
-
-def test_metrics_plot_boxplot_arg_xticklabels_size(mocker):
-    m = mocker.patch("matplotlib.axes.Subplot.set_xticklabels")
-    xticklabels_size = 10
-    data = {"data1": np.array([1, 2, 3, 4]), "data2": np.array([1, 2, 3, 4])}
-
-    bench = jb.Benchmark(
-        params={},
-        solver=[solve],
-    )
-    result = bench()
-    mplot = MetricsPlot(result)
-    mplot.boxplot(
-        f=simple_func_for_boxplot,
-        xticklabels_size=xticklabels_size,
-    )
-
-    m.assert_called_once_with(data.keys(), size=10, rotation=None)
-
-
-def test_metrics_plot_boxplot_arg_xticklabels_rotation(mocker):
-    m = mocker.patch("matplotlib.axes.Subplot.set_xticklabels")
-    xticklabels_rotation = 10
-    data = {"data1": np.array([1, 2, 3, 4]), "data2": np.array([1, 2, 3, 4])}
-
-    bench = jb.Benchmark(
-        params={},
-        solver=[solve],
-    )
-    result = bench()
-    mplot = MetricsPlot(result)
-    mplot.boxplot(
-        f=simple_func_for_boxplot,
-        xticklabels_rotation=xticklabels_rotation,
-    )
-
-    m.assert_called_once_with(data.keys(), size=None, rotation=10)
-
-
-def test_metrics_plot_boxplot_arg_ylabel():
-    ylabel = "ylabel"
-
-    bench = jb.Benchmark(
-        params={},
-        solver=[solve],
-    )
-    result = bench()
-    mplot = MetricsPlot(result)
-    fig_ax_tuple = mplot.boxplot(
-        f=simple_func_for_boxplot,
-        ylabel=ylabel,
-    )
-
-    ax = fig_ax_tuple[0][1]
-    assert ax.get_ylabel() == "ylabel"
-
-
-def test_metrics_plot_boxplot_arg_ylabel_size(mocker):
-    m = mocker.patch("matplotlib.axes.Subplot.set_ylabel")
-
-    ylabel = "ylabel"
-    ylabel_size = 10
-
-    bench = jb.Benchmark(
-        params={},
-        solver=[solve],
-    )
-    result = bench()
-    mplot = MetricsPlot(result)
-    mplot.boxplot(f=simple_func_for_boxplot, ylabel=ylabel, ylabel_size=ylabel_size)
-
-    m.assert_called_once_with("ylabel", size=ylabel_size)
-
-
-def test_metrics_plot_boxplot_arg_yticks_default(mocker):
-    m = mocker.patch("matplotlib.ticker.MaxNLocator.__init__", return_value=None)
-
-    bench = jb.Benchmark(
-        params={},
-        solver=[solve],
-    )
-    result = bench()
-    mplot = MetricsPlot(result)
-    mplot.boxplot(f=simple_func_for_boxplot)
-
-    m.assert_called_with(integer=True)
-
-
-def test_metrics_plot_boxplot_arg_yticks():
-    yticks = [0, 10]
-
-    bench = jb.Benchmark(
-        params={},
-        solver=[solve],
-    )
-    result = bench()
-    mplot = MetricsPlot(result)
-    fig_ax_tuple = mplot.boxplot(
-        f=simple_func_for_boxplot,
-        yticks=yticks,
-    )
-    ax = fig_ax_tuple[0][1]
-    assert (ax.get_yticks() == [0, 10]).all()
-
-
-def test_metrics_plot_boxplot_arg_kwargs(mocker):
-    m = mocker.patch("matplotlib.axes.Subplot.boxplot")
-
-    bench = jb.Benchmark(
-        params={},
-        solver=[solve],
-    )
-    result = bench()
-    mplot = MetricsPlot(result)
-    mplot.boxplot(
-        f=simple_func_for_boxplot,
-        showmeans=True,
-        meanline=True,
-    )
-
-    args, kwargs = m.call_args
-    assert kwargs["showmeans"] is True
-    assert kwargs["meanline"] is True
-
-
 def test_metrics_plot_boxplot_violations_return_value():
     num_multipliers = 2
     params = {"multipliers": [{} for _ in range(num_multipliers)]}
@@ -343,7 +78,7 @@ def test_metrics_plot_boxplot_violations_return_value():
         solver=solver_list,
     )
     result = bench()
-    mplot = MetricsPlot(result)
+    mplot = ConstraintPlot(result)
     fig_ax_tuple = mplot.boxplot_violations()
     assert len(fig_ax_tuple) == expect_num_fig
     assert type(fig_ax_tuple[0][0]) == figure.Figure
@@ -364,7 +99,7 @@ def test_metrics_plot_boxplot_violations_call_maplotlib_boxplot(mocker):
         solver=solver_list,
     )
     result = bench()
-    mplot = MetricsPlot(result)
+    mplot = ConstraintPlot(result)
     mplot.boxplot_violations()
 
     assert m.call_count == expect_num_call
@@ -378,7 +113,7 @@ def test_metrics_plot_boxplot_violations_arg_figsize():
         solver=[solve],
     )
     result = bench()
-    mplot = MetricsPlot(result)
+    mplot = ConstraintPlot(result)
     fig_ax_tuple = mplot.boxplot_violations(
         figsize=(figwidth, figheight),
     )
@@ -402,7 +137,7 @@ def test_metrics_plot_boxplot_violations_arg_title(mocker):
         solver=solver_list,
     )
     result = bench()
-    mplot = MetricsPlot(result)
+    mplot = ConstraintPlot(result)
     mplot.boxplot_violations(
         title=title,
     )
@@ -429,7 +164,7 @@ def test_metrics_plot_boxplot_violations_arg_title_fontsize(mocker):
         solver=solver_list,
     )
     result = bench()
-    mplot = MetricsPlot(result)
+    mplot = ConstraintPlot(result)
     mplot.boxplot_violations(
         title=title,
         title_fontsize=title_fontsize,
@@ -448,7 +183,7 @@ def test_metrics_plot_boxplot_violations_arg_xticklabels():
         solver=[solve],
     )
     result = bench()
-    mplot = MetricsPlot(result)
+    mplot = ConstraintPlot(result)
     fig_ax_tuple = mplot.boxplot_violations()
     ax = fig_ax_tuple[0][1]
     xticklabels = ax.get_xticklabels()
@@ -466,7 +201,7 @@ def test_metrics_plot_boxplot_violations_arg_xticklabels_size(mocker):
         solver=[solve],
     )
     result = bench()
-    mplot = MetricsPlot(result)
+    mplot = ConstraintPlot(result)
     mplot.boxplot_violations(
         constraint_name_fontsize=constraint_name_fontsize,
     )
@@ -483,7 +218,7 @@ def test_metrics_plot_boxplot_violations_arg_xticklabels_rotation(mocker):
         solver=[solve],
     )
     result = bench()
-    mplot = MetricsPlot(result)
+    mplot = ConstraintPlot(result)
     mplot.boxplot_violations(
         constraint_name_fontrotation=constraint_name_fontrotation,
     )
@@ -497,7 +232,7 @@ def test_metrics_plot_boxplot_violations_arg_ylabel_default():
         solver=[solve],
     )
     result = bench()
-    mplot = MetricsPlot(result)
+    mplot = ConstraintPlot(result)
     fig_ax_tuple = mplot.boxplot_violations()
 
     ax = fig_ax_tuple[0][1]
@@ -512,7 +247,7 @@ def test_metrics_plot_boxplot_violations_arg_ylabel():
         solver=[solve],
     )
     result = bench()
-    mplot = MetricsPlot(result)
+    mplot = ConstraintPlot(result)
     fig_ax_tuple = mplot.boxplot_violations(
         ylabel=ylabel,
     )
@@ -532,7 +267,7 @@ def test_metrics_plot_boxplot_violations_arg_ylabel_size(mocker):
         solver=[solve],
     )
     result = bench()
-    mplot = MetricsPlot(result)
+    mplot = ConstraintPlot(result)
     mplot.boxplot_violations(ylabel=ylabel, ylabel_size=ylabel_size)
 
     m.assert_called_once_with("ylabel", size=ylabel_size)
@@ -547,7 +282,7 @@ def test_metrics_plot_boxplot_violations_arg_yticks_default(mocker):
         solver=[solve],
     )
     result = bench()
-    mplot = MetricsPlot(result)
+    mplot = ConstraintPlot(result)
     mplot.boxplot_violations()
 
     m.assert_called_with(integer=True)
@@ -561,7 +296,7 @@ def test_metrics_plot_boxplot_violations_arg_yticks():
         solver=[solve],
     )
     result = bench()
-    mplot = MetricsPlot(result)
+    mplot = ConstraintPlot(result)
     fig_ax_tuple = mplot.boxplot_violations(
         yticks=yticks,
     )
@@ -577,7 +312,7 @@ def test_metrics_plot_boxplot_violations_arg_kwargs_default(mocker):
         solver=[solve],
     )
     result = bench()
-    mplot = MetricsPlot(result)
+    mplot = ConstraintPlot(result)
     mplot.boxplot_violations()
 
     args, kwargs = m.call_args
@@ -593,7 +328,7 @@ def test_metrics_plot_boxplot_violations_arg_kwargs(mocker):
         solver=[solve],
     )
     result = bench()
-    mplot = MetricsPlot(result)
+    mplot = ConstraintPlot(result)
     mplot.boxplot_violations(
         showmeans=True,
         meanline=True,
@@ -618,7 +353,7 @@ def test_metrics_plot_boxplot_violations_hline_indicate_no_violation(mocker):
         solver=solver_list,
     )
     result = bench()
-    mplot = MetricsPlot(result)
+    mplot = ConstraintPlot(result)
     mplot.boxplot_violations()
 
     m.assert_has_calls(
